moved authOptions to separate file due to type issue
